% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAjuste.R
\name{plotAjuste}
\alias{plotAjuste}
\title{Gráfico de Ajuste do Modelo}
\usage{
plotAjuste(modelo = NULL)
}
\arguments{
\item{modelo}{Objeto retornado pela função Gam_Glarma.}
}
\description{
Gera um gráfico comparando os valores reais e os valores ajustados (fitted values) ao longo do tempo.
}
