% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduos.R
\name{plotResiduos}
\alias{plotResiduos}
\title{Diagnóstico de Resíduos}
\usage{
plotResiduos(modelo = NULL)
}
\arguments{
\item{modelo}{Objeto retornado pela função Gam_Glarma.}
}
\description{
Gera gráficos e testes estatísticos (Box-Pierce e Shapiro-Wilk) para os resíduos do modelo, além de recomendar impulsos.
}
