% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gam_Glarma.R
\name{Gam_Glarma}
\alias{Gam_Glarma}
\title{Ajuste de Modelos GAM-GLARMA}
\usage{
Gam_Glarma(
  formula,
  data = NULL,
  offset = NULL,
  type = NULL,
  link = NULL,
  method = "FS",
  residuals = "Score",
  phiLags = NULL,
  thetaLags = NULL,
  phiInit = NULL,
  thetaInit = NULL,
  beta = NULL,
  alphaInit = NULL,
  alpha = 1,
  maxit = 30,
  grad = 2.22e-16,
  n_spline = 0,
  auto_spline = F,
  spline_cols = NULL,
  nos_spline = 3,
  grau_spline = 2,
  sen_cos = c(),
  tendencia = F,
  impulso = 0,
  passo = F,
  n_passos = 0,
  tol_passo = 0.1,
  patience = 3,
  tol_aic_imp = 1e-08,
  trace = FALSE
)
}
\arguments{
\item{formula}{Um objeto da classe \code{formula} (ex: \code{y ~ x1 + x2}).}

\item{data}{Um \code{data.frame} opcional contendo as variáveis.}

\item{type}{Distribuição da família exponencial ("Poisson", "Normal", "Gamma").}

\item{link}{Função de ligação ("log", "identity", "inverse").}

\item{n_spline}{Número de graus de liberdade para as splines.}

\item{auto_spline}{Lógico. Se TRUE, busca o melhor grau de spline (1 a 5) baseado no AIC.}

\item{passo}{Lógico. Se TRUE, utiliza o pacote changepoint para detectar quebras de média.}
}
\value{
Uma lista da classe \code{gamglarma} contendo coeficientes, deviance, AIC e a série latente Z.
}
\description{
GamGlarma é usado para ajustar modelos Glarmas com ou sem partes aditivas, fornecendo os preditores lineares, os coeficientes arma e a descrição dos resíduos.
}
